requirements.txt
# Trading System Requirements
# Install with: pip install -r requirements.txt

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Environment configuration
python-dotenv>=1.0.0

# Alpaca API
alpaca-py>=0.21.0

# Technical Analysis
ta>=0.10.2

# Data Storage
pyarrow>=14.0.0

# Optional: Yahoo Finance for VIX data
yfinance>=0.2.28

# Development/Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-timeout>=2.2.0
pytest-xdist>=3.3.0  # Parallel test execution
coverage>=7.3.0
hypothesis>=6.82.0  # Property-based testing (optional but recommended)

# Research & Optimization
deap>=1.4.1  # Genetic programming for strategy discovery
# scikit-learn>=1.3.0  # Optional: ML models

# Dashboard
flask>=3.0.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0

# Optional: Redis for pub/sub (Phase 3+)
# redis>=5.0.0

# Optional: Jupyter for research
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0